Authentication Check added for map webpage.
Authcheck.js added (lets page check if a user is logged in)
Currently when hitting start and not logged in. Start will take you to -> pagesignin/index.html -> upload.html.

You need to manually go back to the pagemap URL and then in the console you can see that you are signed in with your email.

Stuff Needed:

Add a signifier that the user is logged in.

Take upload.html code and integrate it with the "I found a Slug" button.